Refactor a generic utils module for shared code
### Description
`covidhub_interface_from_secret` is already used in multiple places.  `get_or_make_group` and `retrieve_auth0_users` will be used by `import_users.py` and the future admin import.

#### Issue
[ch135517](https://app.clubhouse.io/genepi/story/135517/port-over-account-info-for-admin-czb-czi-users)

### Test plan
mechanical refactor.
