+++ src/backend/aspen/covidhub_import/__init__.py
@@ -5,4 +5,5 @@ except ImportError:
 else:
     from .import_projects import import_project  # noqa: F401
     from .import_trees import import_trees  # noqa: F401
-    from .import_users import import_project_users, retrieve_auth0_users  # noqa: F401
+    from .import_users import import_project_users  # noqa: F401
+    from .utils import retrieve_auth0_users  # noqa: F401
